requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-telegram-bot==20.3
psutil==5.9.5
requests==2.31.0
